Placed SmashPlant to another module
